investment_portal/requirements.txt
asgiref==3.8.1
Django==4.2.15
django-cors-headers==4.4.0
django-rest-framework==0.1.0
djangorestframework==3.15.2
sqlparse==0.5.1
